Refactor: Add #shit_ciphertext to module Shiftable to handle message decryption
