+++ encrypted.txt
@@ -1,1 +1,1 @@
- mcbg!hnejtv
+gpfcn!kqfqwy
